combine update item test with the rest of item tests
