+++ test/index.js
@@ -3,4 +3,3 @@ require('./test.batch');
 require('./test.convertTypes');
 require('./test.item');
 require('./test.table');
-require('./test.updateItem');
